Automatically handle annotation continuation files.
If PyAnnolib sees files with extensions of "_1", "_2", etc.,
it will "append" them to the first file. It does this via a class
that looks like a file object, but that read across multiple files.

Also, fix a bug where we assume the first emake process in a level 0
make. If emake was started from within a GNU Make process, this is false.